feat(dvr): Finalize and test DVR module logic
