--- reconciliation_reports.log
@@ -1,3 +1,0 @@
-Sensor 1 reconciliation status: OK
-Sensor 3 reconciliation status: BAD
-Sensor 4 reconciliation status: OK
